+++ NEWS.md
@@ -1,3 +1,9 @@
+# EIAapi 0.1.2
+
+* Added a new function, `eia_backfill` to handle large data query from the API
+* Fixed issue with the `eia_get` function - `curl` commands failed to send query with brackets 
+* Added new vignette with example of the `eia_backfill` function usage
+
 # EIAapi 0.1.1
 
 * Modified the `eia_get` function argument - rename the `api_url` argument to `api_path`, and dropping the end-point (`https://api.eia.gov/v2/`) from the query